fix:ТЗ-11 сделано на 1/2 + readme

+++ src/main/java/ru/yandex/practicum/filmorate/storage/LikeStorage.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.storage;
 
 public interface LikeStorage {
-    void createLike(int id, int userId);
+    void createLike(long id, long userId);
 
-    void deleteLike(int id, int userId);
+    void deleteLike(long id, long userId);
 }
